sellerops-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0

pydantic==2.5.3
email-validator

sqlalchemy==2.0.25
psycopg2-binary==2.9.9

celery==5.3.6
redis==5.0.1


python-jose

python-multipart==0.0.9
python-dotenv

pandas
passlib[argon2]

argon2-cffi
alembic==1.13.1

